feat: upgrade gitlab client-go v1; migrate ints to int64; simplify pointer copies

+++ apis/cluster/groups/v1alpha1/zz_samlgrouplink_types.go
@@ -29,7 +29,7 @@ type SamlGroupLinkParameters struct {
 	// GroupID is the ID of the group to create the deploy token in.
 	// +optional
 	// +immutable
-	GroupID *int `json:"groupId,omitempty"`
+	GroupID *int64 `json:"groupId,omitempty"`
 
 	// GroupIDRef is a reference to a group to retrieve its groupId
 	// +optional
@@ -51,7 +51,7 @@ type SamlGroupLinkParameters struct {
 	// memberRoleID is the defined member role assigned to members of the group
 	// +optional
 	// +immutable
-	MemberRoleID *int `json:"memberRoleId,omitempty"`
+	MemberRoleID *int64 `json:"memberRoleId,omitempty"`
 }
 
 // SamlGroupLinkObservation represents a Group Saml Link.
