+++ pkg/cluster/controller/projects/approvalrules/zz_controller_test.go
@@ -39,8 +39,8 @@ import (
 var (
 	unexpecedItem                 resource.Managed
 	errBoom                       = errors.New("boom")
-	projectID                     = 0
-	approvalsRequired             = 1
+	projectID                     = int64(0)
+	approvalsRequired             = int64(1)
 	users                         = []*gitlab.BasicUser{{ID: 123, Username: "abc"}, {ID: 456, Username: "testUser"}}
 	groups                        = []*gitlab.Group{{ID: 99}}
 	protectedBranches             = []*gitlab.ProtectedBranch{{ID: 1}, {ID: 2}}
@@ -184,7 +184,7 @@ func TestObserve(t *testing.T) {
 		"ErrGet404": {
 			args: args{
 				projectApprovalRule: &fake.MockClient{
-					MockGetProjectApprovalRule: func(pid any, ruleID int, options ...gitlab.RequestOptionFunc) (*gitlab.ProjectApprovalRule, *gitlab.Response, error) {
+					MockGetProjectApprovalRule: func(pid any, ruleID int64, options ...gitlab.RequestOptionFunc) (*gitlab.ProjectApprovalRule, *gitlab.Response, error) {
 						return nil, &gitlab.Response{Response: &http.Response{StatusCode: 404}}, errBoom
 					},
 				},
@@ -211,7 +211,7 @@ func TestObserve(t *testing.T) {
 		"ErrGet": {
 			args: args{
 				projectApprovalRule: &fake.MockClient{
-					MockGetProjectApprovalRule: func(pid any, ruleID int, options ...gitlab.RequestOptionFunc) (*gitlab.ProjectApprovalRule, *gitlab.Response, error) {
+					MockGetProjectApprovalRule: func(pid any, ruleID int64, options ...gitlab.RequestOptionFunc) (*gitlab.ProjectApprovalRule, *gitlab.Response, error) {
 						return nil, nil, errBoom
 					},
 				},
@@ -226,7 +226,7 @@ func TestObserve(t *testing.T) {
 		"SuccessfulAvailable": {
 			args: args{
 				projectApprovalRule: &fake.MockClient{
-					MockGetProjectApprovalRule: func(pid any, ruleID int, options ...gitlab.RequestOptionFunc) (*gitlab.ProjectApprovalRule, *gitlab.Response, error) {
+					MockGetProjectApprovalRule: func(pid any, ruleID int64, options ...gitlab.RequestOptionFunc) (*gitlab.ProjectApprovalRule, *gitlab.Response, error) {
 						return &gitlab.ProjectApprovalRule{
 							ApprovalsRequired: approvalsRequired,
 							Name:              name,
@@ -396,7 +396,7 @@ func TestUpdate(t *testing.T) {
 		"SuccessfulUpdate": {
 			args: args{
 				projectApprovalRule: &fake.MockClient{
-					MockUpdateProjectApprovalRule: func(pid any, approvalRule int, opt *gitlab.UpdateProjectLevelRuleOptions, options ...gitlab.RequestOptionFunc) (*gitlab.ProjectApprovalRule, *gitlab.Response, error) {
+					MockUpdateProjectApprovalRule: func(pid any, approvalRule int64, opt *gitlab.UpdateProjectLevelRuleOptions, options ...gitlab.RequestOptionFunc) (*gitlab.ProjectApprovalRule, *gitlab.Response, error) {
 						return &gitlab.ProjectApprovalRule{
 							ID:                            projectID,
 							Name:                          name,
@@ -434,7 +434,7 @@ func TestUpdate(t *testing.T) {
 		"FailedUpdate": {
 			args: args{
 				projectApprovalRule: &fake.MockClient{
-					MockUpdateProjectApprovalRule: func(pid any, approvalRule int, opt *gitlab.UpdateProjectLevelRuleOptions, options ...gitlab.RequestOptionFunc) (*gitlab.ProjectApprovalRule, *gitlab.Response, error) {
+					MockUpdateProjectApprovalRule: func(pid any, approvalRule int64, opt *gitlab.UpdateProjectLevelRuleOptions, options ...gitlab.RequestOptionFunc) (*gitlab.ProjectApprovalRule, *gitlab.Response, error) {
 						return &gitlab.ProjectApprovalRule{}, &gitlab.Response{}, errBoom
 					},
 				},
@@ -490,7 +490,7 @@ func TestDelete(t *testing.T) {
 		"SuccessfulDeletion": {
 			args: args{
 				projectApprovalRule: &fake.MockClient{
-					MockDeleteProjectApprovalRule: func(pid any, approvalRule int, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
+					MockDeleteProjectApprovalRule: func(pid any, approvalRule int64, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
 						return &gitlab.Response{}, nil
 					},
 				},
@@ -517,7 +517,7 @@ func TestDelete(t *testing.T) {
 		"FailedDeletion": {
 			args: args{
 				projectApprovalRule: &fake.MockClient{
-					MockDeleteProjectApprovalRule: func(pid any, approvalRule int, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
+					MockDeleteProjectApprovalRule: func(pid any, approvalRule int64, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
 						return &gitlab.Response{}, errBoom
 					},
 				},
